backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0
pgvector>=0.2.4
pydantic>=2.5.0
pydantic-settings>=2.1.0
anthropic>=0.40.0
openai>=1.12.0
httpx>=0.26.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
numpy>=1.26.0
scikit-learn>=1.4.0
python-multipart>=0.0.6
sse-starlette>=2.0.0
tenacity>=8.2.0
structlog>=24.1.0
